Windows: fix (Get|Set)ThreadContext on Windows x64 these APIs are actually implemented in an async manner, by scheduling APCs to run on the target thread. ezinject was not properly accounting for this, thus risking a race condition and reading garbage values.
while the quickest thing to do would be to put a simple "Sleep" after each invocation,
I decided it would be more beneficial to write a (ugly) workaround that should work in all cases

The workaround hijacks the program counter (which is read/written synchronously by ntoskrnl)
to detour the program to a controlled location, from which we can suspend/resume the thread freely.
This allows us to get/set thread registers and await for the APC completion.