avoid spilling into unsaved stack memory when the stack grows up.
in the previous implementation, `sizeof(*rcall)` was being added to
the stack pointer, and was then used as the new stack base.
this means the first usable location is after the saved stack area,
which is not ideal (we risk a crash, since code will overwrite data
that will not be restored).
we should be using `entry_stack` instead, same as when the stack
grows down.

to account for this change, we need two different trampoline layouts
in the injcode_call structure.
this implies that we need to know the stack direction at compile time.